+++ topic0/fol/AbstractFactory/FactoryProducer.java
@@ -2,11 +2,13 @@ package fol.AbstractFactory;
 
 public class FactoryProducer {
     public static Abstractfactory getFactory(String type) {
-        if (type.equalsIgnoreCase("connection")) {
-            return new ConnectionFactory();
-        }
-        if (type.equalsIgnoreCase("server")) {
-            return new ServerFactory();
+        if(type != "connection" || type !="server") {
+            if (type.equalsIgnoreCase("connection")) {
+                return new DbFactory();
+            }
+            if (type.equalsIgnoreCase("server")) {
+                return new DbFactory();
+            }
         }
         return null;
     }
